requirements.txt
git+https://github.com/Rapptz/discord.py.git
Pillow==9.1.1
xkcd==2.4.2
pytz==2022.1
asyncpg
alembic==1.8.0
python-Levenshtein==0.12.2